test(azure): split responses processor tests
