+++ src/core/providers/azure/responses/processor.rs
@@ -362,525 +362,5 @@ pub struct StreamingChunk {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    // ==================== ResponseProcessingConfig Tests ====================
-
-    #[test]
-    fn test_response_processing_config_default() {
-        let config = ResponseProcessingConfig::default();
-
-        assert!(config.process_content_filters);
-        assert!(config.calculate_metrics);
-        assert!(config.validate_structure);
-        assert_eq!(config.max_response_size, 10 * 1024 * 1024); // 10MB
-    }
-
-    #[test]
-    fn test_response_processing_config_custom() {
-        let config = ResponseProcessingConfig {
-            process_content_filters: false,
-            calculate_metrics: false,
-            validate_structure: false,
-            max_response_size: 1024,
-        };
-
-        assert!(!config.process_content_filters);
-        assert!(!config.calculate_metrics);
-        assert!(!config.validate_structure);
-        assert_eq!(config.max_response_size, 1024);
-    }
-
-    #[test]
-    fn test_response_processing_config_clone() {
-        let config = ResponseProcessingConfig::default();
-        let cloned = config.clone();
-
-        assert_eq!(
-            config.process_content_filters,
-            cloned.process_content_filters
-        );
-        assert_eq!(config.max_response_size, cloned.max_response_size);
-    }
-
-    #[test]
-    fn test_response_processing_config_debug() {
-        let config = ResponseProcessingConfig::default();
-        let debug = format!("{:?}", config);
-        assert!(debug.contains("ResponseProcessingConfig"));
-    }
-
-    // ==================== AzureResponseProcessor Creation Tests ====================
-
-    #[test]
-    fn test_azure_response_processor_new() {
-        let processor = AzureResponseProcessor::new();
-        assert!(processor.config.process_content_filters);
-        assert!(processor.config.calculate_metrics);
-    }
-
-    #[test]
-    fn test_azure_response_processor_default() {
-        let processor = AzureResponseProcessor::default();
-        assert!(processor.config.validate_structure);
-    }
-
-    #[test]
-    fn test_azure_response_processor_with_config() {
-        let config = ResponseProcessingConfig {
-            process_content_filters: false,
-            calculate_metrics: true,
-            validate_structure: false,
-            max_response_size: 5000,
-        };
-
-        let processor = AzureResponseProcessor::with_config(config);
-        assert!(!processor.config.process_content_filters);
-        assert!(processor.config.calculate_metrics);
-        assert!(!processor.config.validate_structure);
-        assert_eq!(processor.config.max_response_size, 5000);
-    }
-
-    // ==================== Process Response Tests ====================
-
-    #[test]
-    fn test_process_response() {
-        let processor = AzureResponseProcessor::new();
-        let response = serde_json::json!({
-            "choices": [{"message": {"content": "test"}, "finish_reason": "stop"}],
-            "usage": {"total_tokens": 10}
-        });
-
-        let result = processor.process_response(response).unwrap();
-        assert!(!result.content_filtered);
-    }
-
-    #[test]
-    fn test_process_response_with_id() {
-        let processor = AzureResponseProcessor::new();
-        let response = serde_json::json!({
-            "id": "chatcmpl-123456",
-            "model": "gpt-4",
-            "choices": [{"message": {"content": "Hello"}, "finish_reason": "stop"}],
-            "usage": {"total_tokens": 15}
-        });
-
-        let result = processor.process_response(response).unwrap();
-        assert!(result.metadata.request_id.is_some());
-        assert_eq!(result.metadata.request_id.unwrap(), "chatcmpl-123456");
-        assert!(result.metadata.deployment_id.is_some());
-        assert_eq!(result.metadata.deployment_id.unwrap(), "gpt-4");
-    }
-
-    #[test]
-    fn test_process_response_without_usage_warning() {
-        let processor = AzureResponseProcessor::new();
-        let response = serde_json::json!({
-            "choices": [{"message": {"content": "test"}, "finish_reason": "stop"}]
-        });
-
-        let result = processor.process_response(response).unwrap();
-        assert!(result.warnings.iter().any(|w| w.contains("missing usage")));
-    }
-
-    #[test]
-    fn test_process_response_content_filtered() {
-        let processor = AzureResponseProcessor::new();
-        let response = serde_json::json!({
-            "choices": [{"message": {"content": ""}, "finish_reason": "content_filter"}],
-            "usage": {"total_tokens": 5}
-        });
-
-        let result = processor.process_response(response).unwrap();
-        assert!(result.content_filtered);
-        assert!(result.warnings.iter().any(|w| w.contains("filtered")));
-    }
-
-    #[test]
-    fn test_process_response_exceeds_size_limit() {
-        let config = ResponseProcessingConfig {
-            max_response_size: 10, // Very small limit
-            ..Default::default()
-        };
-        let processor = AzureResponseProcessor::with_config(config);
-
-        let response = serde_json::json!({
-            "choices": [{"message": {"content": "This is a long response"}, "finish_reason": "stop"}]
-        });
-
-        let result = processor.process_response(response);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("exceeds limit"));
-    }
-
-    #[test]
-    fn test_process_response_with_metrics_disabled() {
-        let config = ResponseProcessingConfig {
-            calculate_metrics: false,
-            ..Default::default()
-        };
-        let processor = AzureResponseProcessor::with_config(config);
-
-        let response = serde_json::json!({
-            "choices": [{"message": {"content": "test"}, "finish_reason": "stop"}],
-            "usage": {"total_tokens": 10}
-        });
-
-        let result = processor.process_response(response).unwrap();
-        assert_eq!(result.metrics.total_time_ms, 0);
-    }
-
-    #[test]
-    fn test_process_response_with_validation_disabled() {
-        let config = ResponseProcessingConfig {
-            validate_structure: false,
-            ..Default::default()
-        };
-        let processor = AzureResponseProcessor::with_config(config);
-
-        // Invalid structure that would fail validation
-        let response = serde_json::json!({
-            "choices": []
-        });
-
-        // Should succeed because validation is disabled
-        let result = processor.process_response(response);
-        assert!(result.is_ok());
-    }
-
-    // ==================== Validate Structure Tests ====================
-
-    #[test]
-    fn test_validate_chat_structure() {
-        let processor = AzureResponseProcessor::new();
-        let response = serde_json::json!({
-            "choices": [{"message": {"content": "test"}, "finish_reason": "stop"}]
-        });
-
-        assert!(processor.validate_response_structure(&response).is_ok());
-    }
-
-    #[test]
-    fn test_validate_chat_structure_with_text() {
-        let processor = AzureResponseProcessor::new();
-        let response = serde_json::json!({
-            "choices": [{"text": "completion text", "finish_reason": "stop"}]
-        });
-
-        assert!(processor.validate_response_structure(&response).is_ok());
-    }
-
-    #[test]
-    fn test_validate_chat_structure_empty_choices() {
-        let processor = AzureResponseProcessor::new();
-        let response = serde_json::json!({
-            "choices": []
-        });
-
-        let result = processor.validate_response_structure(&response);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("Empty choices"));
-    }
-
-    #[test]
-    fn test_validate_chat_structure_missing_content() {
-        let processor = AzureResponseProcessor::new();
-        let response = serde_json::json!({
-            "choices": [{"finish_reason": "stop"}]
-        });
-
-        let result = processor.validate_response_structure(&response);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("missing message or text"));
-    }
-
-    #[test]
-    fn test_validate_chat_structure_missing_finish_reason() {
-        let processor = AzureResponseProcessor::new();
-        let response = serde_json::json!({
-            "choices": [{"message": {"content": "test"}}]
-        });
-
-        let result = processor.validate_response_structure(&response);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("missing finish_reason"));
-    }
-
-    #[test]
-    fn test_validate_embedding_structure() {
-        let processor = AzureResponseProcessor::new();
-        let response = serde_json::json!({
-            "data": [{"embedding": [0.1, 0.2, 0.3], "index": 0}],
-            "model": "text-embedding-ada-002"
-        });
-
-        assert!(processor.validate_response_structure(&response).is_ok());
-    }
-
-    #[test]
-    fn test_validate_embedding_structure_empty_data() {
-        let processor = AzureResponseProcessor::new();
-        let response = serde_json::json!({
-            "data": []
-        });
-
-        let result = processor.validate_response_structure(&response);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("Empty embedding"));
-    }
-
-    #[test]
-    fn test_validate_embedding_structure_missing_embedding() {
-        let processor = AzureResponseProcessor::new();
-        let response = serde_json::json!({
-            "data": [{"index": 0}]
-        });
-
-        let result = processor.validate_response_structure(&response);
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("missing embedding field"));
-    }
-
-    #[test]
-    fn test_validate_image_generation_structure() {
-        let processor = AzureResponseProcessor::new();
-        // Note: The validation logic checks for embedding data first if "data" is present
-        // Image generation with "data" array falls through embedding validation
-        // which requires "embedding" field. This is expected behavior - image response
-        // validation needs a non-empty data array with any structure
-        let response = serde_json::json!({
-            "created": 1700000000,
-            "data": [{"embedding": [0.1], "url": "https://example.com/image.png"}]
-        });
-
-        assert!(processor.validate_response_structure(&response).is_ok());
-    }
-
-    #[test]
-    fn test_validate_image_generation_structure_empty_data() {
-        let processor = AzureResponseProcessor::new();
-        // Empty data triggers embedding validation path which checks for non-empty array
-        let response = serde_json::json!({
-            "created": 1700000000,
-            "data": []
-        });
-
-        let result = processor.validate_response_structure(&response);
-        assert!(result.is_err());
-        // Empty data triggers "Empty embedding" error since data validation comes first
-        assert!(result.unwrap_err().contains("Empty embedding"));
-    }
-
-    #[test]
-    fn test_validate_unknown_structure() {
-        let processor = AzureResponseProcessor::new();
-        // Unknown structure should pass validation
-        let response = serde_json::json!({
-            "unknown": "data"
-        });
-
-        assert!(processor.validate_response_structure(&response).is_ok());
-    }
-
-    // ==================== Content Filtering Tests ====================
-
-    #[test]
-    fn test_check_content_filtering_none() {
-        let processor = AzureResponseProcessor::new();
-        let response = serde_json::json!({
-            "choices": [{"message": {"content": "Hello"}, "finish_reason": "stop"}]
-        });
-
-        assert!(!processor.check_content_filtering(&response));
-    }
-
-    #[test]
-    fn test_check_content_filtering_by_finish_reason() {
-        let processor = AzureResponseProcessor::new();
-        let response = serde_json::json!({
-            "choices": [{"message": {"content": ""}, "finish_reason": "content_filter"}]
-        });
-
-        assert!(processor.check_content_filtering(&response));
-    }
-
-    #[test]
-    fn test_check_content_filtering_multiple_choices() {
-        let processor = AzureResponseProcessor::new();
-        let response = serde_json::json!({
-            "choices": [
-                {"message": {"content": "Hello"}, "finish_reason": "stop"},
-                {"message": {"content": ""}, "finish_reason": "content_filter"}
-            ]
-        });
-
-        assert!(processor.check_content_filtering(&response));
-    }
-
-    // ==================== Streaming Chunk Tests ====================
-
-    #[test]
-    fn test_process_streaming_chunk_normal() {
-        let processor = AzureResponseProcessor::new();
-        let chunk = serde_json::json!({
-            "choices": [{"delta": {"content": "Hello"}, "finish_reason": null}]
-        });
-
-        let result = processor.process_streaming_chunk(chunk, false).unwrap();
-        assert!(!result.is_final);
-        assert!(!result.content_filtered);
-        assert!(result.warnings.is_empty());
-    }
-
-    #[test]
-    fn test_process_streaming_chunk_final() {
-        let processor = AzureResponseProcessor::new();
-        let chunk = serde_json::json!({
-            "choices": [{"delta": {}, "finish_reason": "stop"}]
-        });
-
-        let result = processor.process_streaming_chunk(chunk, true).unwrap();
-        assert!(result.is_final);
-    }
-
-    #[test]
-    fn test_process_streaming_chunk_filtered() {
-        let processor = AzureResponseProcessor::new();
-        let chunk = serde_json::json!({
-            "choices": [{"delta": {}, "finish_reason": "content_filter"}]
-        });
-
-        let result = processor.process_streaming_chunk(chunk, true).unwrap();
-        assert!(result.content_filtered);
-        assert!(result.warnings.iter().any(|w| w.contains("filtered")));
-    }
-
-    // ==================== StreamingChunk Tests ====================
-
-    #[test]
-    fn test_streaming_chunk_creation() {
-        let chunk = StreamingChunk {
-            data: serde_json::json!({"test": "data"}),
-            is_final: false,
-            content_filtered: false,
-            warnings: vec![],
-        };
-
-        assert!(!chunk.is_final);
-        assert!(!chunk.content_filtered);
-        assert!(chunk.warnings.is_empty());
-    }
-
-    #[test]
-    fn test_streaming_chunk_with_warnings() {
-        let chunk = StreamingChunk {
-            data: serde_json::json!({}),
-            is_final: true,
-            content_filtered: true,
-            warnings: vec!["Warning 1".to_string(), "Warning 2".to_string()],
-        };
-
-        assert!(chunk.is_final);
-        assert!(chunk.content_filtered);
-        assert_eq!(chunk.warnings.len(), 2);
-    }
-
-    #[test]
-    fn test_streaming_chunk_clone() {
-        let chunk = StreamingChunk {
-            data: serde_json::json!({"content": "test"}),
-            is_final: false,
-            content_filtered: false,
-            warnings: vec!["test warning".to_string()],
-        };
-
-        let cloned = chunk.clone();
-        assert_eq!(chunk.is_final, cloned.is_final);
-        assert_eq!(chunk.warnings, cloned.warnings);
-    }
-
-    #[test]
-    fn test_streaming_chunk_debug() {
-        let chunk = StreamingChunk {
-            data: serde_json::json!({}),
-            is_final: false,
-            content_filtered: false,
-            warnings: vec![],
-        };
-
-        let debug = format!("{:?}", chunk);
-        assert!(debug.contains("StreamingChunk"));
-    }
-
-    // ==================== Metrics Tests ====================
-
-    #[test]
-    fn test_calculate_metrics() {
-        let processor = AzureResponseProcessor::new();
-        let response = serde_json::json!({
-            "choices": [{"message": {"content": "test"}, "finish_reason": "stop"}],
-            "usage": {"total_tokens": 10}
-        });
-
-        let result = processor.process_response(response).unwrap();
-        // Metrics should be populated
-        assert!(result.metrics.response_size_bytes > 0);
-    }
-
-    // ==================== Warning Collection Tests ====================
-
-    #[test]
-    fn test_collect_warnings_empty_choices() {
-        let config = ResponseProcessingConfig {
-            validate_structure: false,
-            ..Default::default()
-        };
-        let processor = AzureResponseProcessor::with_config(config);
-
-        let response = serde_json::json!({
-            "choices": []
-        });
-
-        let warnings = processor.collect_warnings(&response);
-        assert!(warnings.iter().any(|w| w.contains("empty choices")));
-    }
-
-    #[test]
-    fn test_collect_warnings_no_issues() {
-        let processor = AzureResponseProcessor::new();
-        let response = serde_json::json!({
-            "choices": [{"message": {"content": "test"}, "finish_reason": "stop"}],
-            "usage": {"total_tokens": 10}
-        });
-
-        let warnings = processor.collect_warnings(&response);
-        assert!(warnings.is_empty());
-    }
-
-    // ==================== Metadata Extraction Tests ====================
-
-    #[test]
-    fn test_extract_metadata_with_model() {
-        let processor = AzureResponseProcessor::new();
-        let response = serde_json::json!({
-            "model": "gpt-4-turbo",
-            "id": "resp-123",
-            "choices": [{"message": {"content": "test"}, "finish_reason": "stop"}]
-        });
-
-        let metadata = processor.extract_metadata(&response);
-        assert_eq!(metadata.deployment_id, Some("gpt-4-turbo".to_string()));
-        assert_eq!(metadata.request_id, Some("resp-123".to_string()));
-    }
-
-    #[test]
-    fn test_extract_metadata_empty() {
-        let processor = AzureResponseProcessor::new();
-        let response = serde_json::json!({});
-
-        let metadata = processor.extract_metadata(&response);
-        assert!(metadata.deployment_id.is_none());
-        assert!(metadata.request_id.is_none());
-    }
-}
+#[path = "processor_tests.rs"]
+mod tests;
